Move the ripple animation to helper method
